Move minDifficulty and powValue calculations to BlockHeader

+++ ethereumj-core/src/main/java/org/ethereum/config/Constants.java
@@ -1,5 +1,7 @@
 package org.ethereum.config;
 
+import java.math.BigInteger;
+
 public class Constants {
 
     public static int GENESIS_DIFFICULTY = 131072;
@@ -9,7 +11,7 @@ public class Constants {
     public static int MIN_GAS_LIMIT = 125000;
     public static int GAS_LIMIT_BOUND_DIVISOR = 1024;
     public static int MINIMUM_DIFFICULTY = 131072;
-    public static int DIFFICULTY_BOUND_DIVISOR = 2048;
+    public static BigInteger DIFFICULTY_BOUND_DIVISOR = BigInteger.valueOf(2048);
     public static int DURATION_LIMIT = 8;
 
     public static int UNCLE_GENERATION_LIMIT = 7;
